WIP rename shared local arena to physics arena
