+++ server/src/movement.rs
@@ -1,1 +1,1 @@
-pub use vibe_land_shared::local_arena::{MoveConfig, PhysicsArena, Vec3};
+pub use vibe_land_shared::physics_arena::{MoveConfig, PhysicsArena, Vec3};
